[feat/#69]: 커스텀 훅으로 로직 분리
