+++ components/agits/AgitHeader.module.css
@@ -1,0 +1,9 @@
+.agit_filter {
+  min-height: 60px;
+  padding: 14px 20px;
+  box-sizing: border-box;
+}
+.agit_filter .skeleton {
+  width: 100px;
+  height: 32px;
+}
