+++ hooks/index.js
@@ -1,6 +1,7 @@
+import useCallAgitInfo from './useCallAgitInfo';
 import useModal from './useModal';
 import useNavVisible from './useNavVisible';
 import useRedirect from './useRedirect';
 import useToast from './useToast';
 
-export { useModal, useNavVisible, useRedirect, useToast };
+export { useCallAgitInfo, useModal, useNavVisible, useRedirect, useToast };
